Fix post-initialization usage of .datepicker()
After a datepicker has been instantiated with .datepicker({...}),
.datepicker(...) may still be called with other, overloaded signatures.
Since we only want to intercept those times that it is called to
instantiate a new object, we need to inspect `arguments`.

In addition, we need to pass `arguments` on to original function, rather
than [opts], since there may be more than one argument.