Add bug fix for datepicker widget dropdown location
There is an issue in jquery-ui with the datepicker dropdown appearing on
the wrong parts of the screen. There is a pull request to the jquery-ui
branch to fix it, but it is still awaiting approval. This takes that fix
and applies legacy 1.11.4 branch currently used for the 508-plugins. The
jquery-ui.js and jquery-ui.min.js have been updated with the bugfix and
built with the download.jqueryui.com builder locally. The css files also
have been built with that and updated here, but the only change is the
date generated in the commented documentation on the file.